Feature: fencing: add support for operation phases to device properties
This adds support for having multiple "phases" of an operation in the
device_properties_t structure. This will be used to support remapping
sequential topology reboots to all-off-then-all-on.

This is not complete support, only support within the one structure and
its associated code, so at this point, phase is (still) always 0 and these
changes have no effect.