Feature: fencing: support remapping reboots in fencing query results
If a topology fencing request has a level with multiple devices, a requested
reboot should be remapped to all-off-then-all-on (consider the case of remote
power switches for redundant power supplies).

This revision implements the first step in that process. When stonithd
peers receive a fence query for a "reboot" action, they will respond with the
action-specific timeout etc. for "reboot", "off" and "on" instead of just
"reboot".

Because a peer may be allowed to execute some but not all of the actions,
this also introduces a new XML attribute in the query result indicating
if an action is disallowed.

The originating stonithd does not do anything special with the results yet,
so no actual remapping is done.