Feature: fencing: support remapping reboots when processing fencing query results
If a topology fencing request has a level with multiple devices, a requested
reboot should be remapped to all-off-then-all-on (consider the case of remote
power switches for redundant power supplies).

This revision implements the second step in that process. When the originating
stonithd receives all peers' replies to a "reboot" query, it will parse values
for "off" and "on" as well as "reboot".

These values include the new XML attribute indicating whether the peer
is only allowed to run one of the actions and not another.

The originating stonithd does not do anything special with the values yet,
so no actual remapping is done.