Fix: crmd: Do not enforce restart for resources that support reload and don't support any unique parameters
Previously, for a resource that supports reload, if it doesn't support
any unique parameters, restart will be enforced when parameters change,
although apparently they are all non-unique parameters.

With this fix, as long as the resource supports reload,
"op-force-restart" and "op-restart-digest" will be added into lrm_rsc_op
to indicate that, no matter if the resource actually supports any
parameters with unique="1".

This also fixes memory leak in append_restart_list().