Feature: fencing: add ability to target by attribute
Previously, fencing topologies levels could be registered only by node name
regex. The level would apply to any nodes whose names matched the regex.

Now, nodes may also be targeted by the presence of a node attribute. If a
topology level's target is specified as an ATTR=VALUE string, then stonithd
will check for the specified node attribute in its locally cached copy of the
CIB when deciding whether the level applies to a node.

Because this uses the CIB, it may be used only with permanent node attributes,
and may not be used in stonithd's standalone mode.