doc: Pacemaker Remote: replace LXC references
The old LXC walk-through is no longer valid, and running pacemaker_remote
inside a container is no longer considered the preferred means of integrating
containers with Pacemaker, now that the "isolated resources" functionality
is available.

For now at least, replace all LXC references in Pacemaker Remote with a
small note at the end of the KVM walk-through.