pacemaker_remote: set LSB Provides header to the service name
Originally it was set to the daemon name (pacemaker_remoted), but that's
out of line with the script name, the systemd service name and the general
trend, which the main Pacemaker daemon follows as well.

Differing names also break the automatic synchronization of SysV init and
systemd service statuses (as done in Debian at least).

Signed-off-by: Ferenc Wágner <wferi@niif.hu>