Fix: liblrmd: make LSB header parsing more efficient
Previously, some helper code containing a "continue" statement was made into a
macro encapsulated in "do...while(0)", thus making the "continue" ineffective.
Now, it's an inline function, and the continue is done in the caller.

The end result is the same, but the loop is short-circuited when appropriate
as originally intended, and it makes coverity happy.