Fix: crmd: properly trigger follow-up for remote node status changes
6ce4a5e modified process_remote_node_action() to check for failed
remote nodes as well as newly reintegrated ones, however it got the sense of a
boolean test wrong, rendering the function largely ineffective.

This should be a complete fix for failed remote node handling, but it does not
fix reintegration handling, as there is an unrelated problem with that that
pre-existed 6ce4a5e.