Medium: use crm_snprintf_offset to prevent buffer overflow
As we are now detecting "generated string won't fit the buffer", there
is no need to partial (=buggy) restrictions on variable parts of the
string being built.

Also include respective header file when the direct dependency
introduced with crm_snprintf_offset use and that file is not
directly included yet.

This also removes usage of macro attr_snprintf introduced as an interim
solution exclusively in lib/cib/cib_attrs.c.  It was facing a criticism
that it is not so intuitive as offset += MACRO(...) notation and it
wasn't so versatile when it came to error detection, anyway.

Most parts handled semiautomatically thanks to applying (variations of)
semantic patches with (fantastic) coccinelle/spatch tool;
some versions are available:
https://fedorapeople.org/cgit/jpokorny/public_git/coccinelle-patches.git/tree/pacemaker

Signed-off-by: Jan Pokorný <jpokorny@redhat.com>