Feature: libcluster: set remote peer cache state based on CIB node state
crm_remote_peer_cache_refresh() repopulates the remote peer cache based
on the CIB. Previously, it would wipe the cache first, and always assume
the node was online. Now, it updates the cache in place, and sets node state
based on the value of XML_NODE_IN_CLUSTER ("in_ccm") in the CIB node state
(or leaves the state unknown if the remote node is in the CIB configuration
but does not have a CIB node state entry).

Currently, the CIB node state is always member, so this has no real effect,
but that will change later.