Fix: systemd: Return PCMK_OCF_UNKNOWN_ERROR instead of PCMK_OCF_NOT_INSTALLED for uncertain errors on LoadUnit
Previously, as long as we could not get the path of the unit via
LoadUnit, it would return PCMK_OCF_NOT_INSTALLED, which is a hard error.

It should not return PCMK_OCF_NOT_INSTALLED unless it's sure about that.
For uncertain errors such as "org.freedesktop.DBus.Error.NoReply", it
should return PCMK_OCF_UNKNOWN_ERROR instead.