Use the clusterlabs Publican brand for building the books
It's possible to use the in-tree "clusterlabs" brand for building the books
without installing the brand, which requires root access and has permanent
effect on the system.  This patch achieves this without changing the source,
thus the web versions shouldn't be affected, but the price is using the
../xsl symlink a directory level higher than the Makefile, which might be
considered unexpected.

Signed-off-by: Ferenc Wágner <wferi@niif.hu>