Mid : Addition of the reconnection option for Attrd clients.
When a client stops, the use and the client may fail in the update of
the attribute by attrd_update_delegate function.

This is because IPC connection is recycled in attrd_update_delegate
function.
This correction adds an option and evades a problem.

We use attrd_update_delegate function from tools of the Japanese
community, but do not work by this problem well now.
* https://github.com/linux-ha-japan/pm_extras

Best Regards,
Hideo Yamauchi.