Feature: pacemaker_remote: support graceful stops
When pacemaker_remote gets an interrupt signal, if there are any connected
proxy providers, it will send an lrmd IPC op for a shutdown request.
If the provider acknowledges the request, pacemaker_remote will stop accepting
new provider connections and wait until all providers disconnect before
exiting itself. This gives the cluster the opportunity to stop any resources
running on the node that is shutting down.

If the provider is an older version that does not support graceful stops,
pacemaker_remote will time out waiting for the ack, then exit immediately.

Since we are now waiting for resources to exit, the systemd stop timeout
for pacemaker_remote has been raised to match pacemaker's.