Fix: header == NULL when parsing compressed message
When parsing a compressed message, message header is reloaded from the
uncompressed buffer. However, because remote->buffer_offset is zeroed
prior that, it'll consider the buffer already processed and return NULL.
Zeoring remote->buffer_offset after header retrieval should remedy this.

AFAICS, the decompression branch is not used so this perhaps is not an
issue in the real world.