Fix: attrd: ensure remote nodes are in correct peer cache
If attrd receives an update for an unknown node name, it assumes the unknown
node is a cluster node, and adds it to the cluster peer cache.

Previously, if the name was later used for a remote node, that would prevent
its attributes from being written to the CIB. Now, when an attribute is
received for a remote node, attrd will purge any inactive cluster peer cache
entry before adding the node to the remote peer cache.