Feature: Allow cluster and remote LRM API versions to diverge
Requiring the same version is not realistic.

Common examples will be rolling upgrades (where typically you'd upgrade
the cluster first) and guest introspection (where you may never upgrade
the guest software).

Normally you want the "server" to be the higher, however the
client/sever roles are turned on their head here since the cluster is
acting as the "client" connecting to pacemaker remote "server".

However we generally want pacemaker-remote to be as dumb as possible
which is consistent with how we'd do upgrades. So for this reason we
require that the remote node's API version be less-than or equal to that
of the cluster.