xml: make <status> optional in all the RNG schemas
The main reason behind this is that the only materialized, local CIB
snapshot when pacemaker is not running is /var/lib/pacemaker/cib/cib.xml
(and possibly siblings), i.e., a file that is privately managed in due
course of its proceeding, and for storing that file, the CIB content is
first massaged in lib/cib/cib_file.c:cib_file_prepare_xml with
a following justification:

    /* Delete status section before writing to file, because
     * we discard it on startup anyway, and users get confused by it */

This means that status is omitted so as not to cause confusion,
should it be investigated by anyone.  This revision makes the schemas
(admittedly, also retroactively, which should be exceptional for good
reason) reflect this special handling of persisted files by making all
status section optional, effectively ensuring their proper validity.