Fix: pengine: properly order stop actions relative to stonith
This corrects a couple of issues with ordering stop actions relative to a
fence event, related to the distinction between the stonith target and the
action node. Before guest nodes, those were always the same; but now,
a stop action on a guest node may be implied by a fencing of its host,
so they may be different.

1. The log messages sometimes referred to the action node,
when the stonith target was actually meant.

2. Some code intended to break certain loops had become dead code
once the stonith_up event was no longer used and nodes to be fenced
were always marked as unclean. But once guest nodes were possible,
stop actions on a clean guest node could inadvertently (and mistakenly)
trigger the not-so-dead code.

This also improves comments and log messages a bit.