Low: ping resource: avoid undefined func w/ OCF_FUNCTIONS=/dev/null
When generating man pages for the agents:
pacemaker/extra/resources/ping: line 357: ocf_is_true: command not found

Note that neither set of "true" indicators (those in ocf_is_true
function of ocf-shellfuncs[*] file vs. those recognized as values
of "debug" OCF parameter in the very script) is a subset of the other,
but acknowledge that latter is a master on its own domain
(at least for now, also to avoid regression when "True" is used).

[*] as an aside, there's also an almost anecdotical "ja" Germanism