Feature: pengine: add meta-attribute to stops for recovery attempts remaining
If a stop action is part of a resource recovery on a node, a new meta-attribute
will be added with the number of additional failures that can happen before
the resource is forced away from the node. The resource agents will get this
value as the OCF_RESKEY_CRM_meta_restart_remaining environment variable.

If a container filler fails, the current implementation adds the variable to
the container stop rather than the filler stop. While this is not ideal,
it should be irrelevant in practice, as fillers are only used for
guest nodes' Pacemaker Remote connections (which don't care about the
new variable) and nagios resources (which ignore stop actions).