Build: spec: avoid user-triggered cluster.conf validation failure
...when run in a bad moment, i.e., when having configuration that the
Pacemaker+CMAN cluster will be started with in cluster.conf statically
(provided that this configuration refers to fence_pcmk fence agent,
which it should as fencing it an imperative, to restate the obvious).
This statement missing is likely a reason why pcs, when setting up new
Pacemaker+CMAN cluster, resorted to invoking ccs with "-i" switch so
as to avoid validation and its imminent failure in a scenario like that.

See also:
<https://bugzilla.redhat.com/1116401#Postinstall-scriptlet-is-missing-ccs_update_schema-invocation>