st_client: print the full unexpected status value from waitpid()
Neither WIFSTOPPED nor WIFCONTINUED could evaluate to true here, because
the waitpid() calls above lack the WUNTRACED and WCONTINUED flags.  Since
we don't use ptrace() this is an "impossible" branch, but it's best to
have the raw status value around to enable post-mortem manual decoding.

This patch also enhances portability, because WIFCONTINUED is not defined
on Hurd (at least).

Signed-off-by: Ferenc Wágner <wferi@niif.hu>