Fix: crmd: Acknowledge cancellation operations for remote connection resources
Previously, putting a remote connection resource into maintenance mode
would cause crmd hanging on waiting for the cancellation of the monitor
operation to return until it timed out.

The cancellation operation for a remote connection resource is
synchronous. We should acknowledge it directly.

It also removes the pending operation with the correct operation id.