Feature: pengine: add meta-attribute when start in same transition
If a resource has demote and/or stop actions followed by a start action in the
same transition, add a new meta-attribute to indicate this. Resource agents
will see this as the OCF_RESKEY_CRM_meta_start_expected environment variable,
with value "local" if start is expected on same node, and "peer" if start is
expected on another node.