Feature: pengine: add meta-attribute to indicate restart attempts
If a resource has a stop action followed by a start action on the same node in
the same transition, a new meta-attribute will be added to indicate this.
Resource agents will see this as the OCF_RESKEY_CRM_meta_restarting=true
environment variable, allowing them to do a "lightweight" stop if desired.