Fix: pengine: avoid use-after-free with location constraint + sets + templates
Previously, rsc2node_new() assumed id would be persistent, but when a location
constraint involves a resource set referencing a template, unpack_location()
will use a copy of the XML and free it afterward. Now, it makes a copy of id,
and pe_free_rsc_to_node() will free that copy when appropriate.