alerts: fix SMTP recipient defaulting
The ${parameter=word} construct tests for unset parameter (the := variant
tests for unset or null), but email_recipient can't be unset after its
initial assignment even if CRM_alert_recipient was unset.  Simply adding
a colon would be enough but not very transparent, so let's merge the initial
assignment into the defaulting.

Signed-off-by: Ferenc Wágner <wferi@niif.hu>