xml: score-attribute{,-mangle} has no meaning for rsc_colocation
...and never had.

This was introduced as of f722c9f commit and already reflected to-be
committed change (the next day) 8c7e2b5 that turned underscores to
dashes for some XML-relevant code-tracked elements/attributes, and
there, in turn, the title pair of labels was introduced as of fbc1e24.

Likely by mistake, there was a missing 'd' in RNG-tracked version
of score-attribute-mangled, and furthermore it seems that rsc_colocation
should not have, RNG-wise,received the parameters per title, unlike
the rule element, which actually did receive just score-attribute
within the same commit introductory commit (f722c9f).  Rationale for
that is that as of ea1359b, the discussed in-code label got finally
used, but rather in the context of rsc_location's rule.

As we can conclude that attaching both score-attribute and
score-attribute-mangle to rsc_colocation in RNG schemas was rather
accidental from the beginning (not to speak about the mismatch
in the exact form of the latter between schema and the code),
and it would be meaningless trying to use that to achieve anything,
the correction is, as an exception, performed retrospectively all
the way down through respective versions.