Low: ping: Avoid temp files in fping_check  (bsc#987348)
Unfortunately the previous attempt to avoid this issue (commit
7eb958ce4) didn't work since the problem happens only when the
output buffer is flushed but the write fails. Since there is no
mechanism in the shell to check if a pipe write is successful,
this silently fails.

In ping_check, there are no temporary files created. It's
possible to avoid the temporaries for fping too, but I had to
resort to using python.