Low: schemas.c: fix false positive of "ret ver < orig ver" indicator
...in update_validation with transform==FALSE and when not any following
schema (we do not decide for the current one in that case) can validate
the input CIB as otherwise "ret ver" (*best) would be almost certainly
returned as a value less than original version (i.e., default 0, while
the original is extremely likely higher).

This wasn't an issue in the past, but it might be in the future since
commit e4b9f34 promoted such inequality into "fancy nonconforming CIB"
indication -- note that cli_config_update itself calls update_validation
with transform==TRUE, so it's out of this question.