Doc: overhaul the README and other top-level documentation
NEWS was never useful. TODO had gotten woefully out of date
and is no longer used by core developers to manage to-do items.
Both were removed.

AUTHORS was also removed and replaced by a Q&A in README.

README was split into README, CONTRIBUTING and INSTALL.
A separate CONTRIBUTING file is useful because github will
display a link to it when someone submits a pull request.