Build: PKG_CHECK_VAR to get corosync path for plugins
Amongst other advantages, this allows for user overrides via
LCRSODIR environment variable; this supplies what --with-lcrso-dir=DIR
option to configure was meant to offer (if only it hadn't been broken
and thus removed in the previous commit).

This also avoids emitting a line like this in the configure output:

> Corosync Plugins         = /usr/lib

in case pacemaker is not configured with corosync plugin architecture
in mind.