Build: spec: reorganize so %pre etc. blocks are self-contained
Instead of one big "if systemd" wrapping all the blocks,
put "if systemd" within each block.

It is debatable whether this is more or less readable, but the advantage is
that someone looking at a small part of the file or a diff can understand
what's going on with a minimum of context, rather than have to scroll back
and forth for many pages of context.