Build: configure: do not support "prefix" in --with-initdir value
Relative prefix-rooted path doesn't make much sense for standard init
systems' paths, and one can still expand the prefix prior to passing
the value via the mentioned switch.

Also, do not try to be overly smart when detecting correct location
for init scripts based on what exists when not explicitly specified.
That's what --with-initdir switch is for, without a risk of false
positives provided that who initiates the build knows the target
system well enough.