Build: spec: reflect init system + init launchers exclusivity
This commit serves to adapt spec file to the previous commit that
introduced controlled way to configure where to install pacemaker
init launchers pertaining the main init system as selected/detected.

The arrangement in the specfile actually overrides the inner build
system decision about which init system to stick with, because:
. at the time of running %prep scriptlet, it's already known which
  init system is used as the main one (by order of priority):
  - _unitdir RPM macro defined -> systemd
  - user builds using --with-upstart -> Upstart
  - otherwise -> LSB
. appropriate --with-initdir=DIR to ./configure can only be chosen
  when init system is clear

Practically, this enforces init launchers to always be installed, but
that's actually expected.  On the bright side, it allows ugly
on-top-of-make-install messing with files that should really be
part of a proper installation process (and what's more, the
infrastructure is already there as of the previous commit!) to
be dropped.

This also allows for conditionalizing on pkgconfig(dbus-1) BR.