Build: spec: more robust check for systemd being in use
The spec file parsing prior to having even BuildRequires installed
is pretty crude (e.g. when RPM building system is just to evaluate
these BRs, so apparently, they cannot be magically pre-existed).
At least when building in mock, as there's hardly anything that would
give a hint about the target system beside distribution-defined macros
(fedora, rhel, ...).

So also reflect this in the comment conditionalizing on bleeding macro,
which is driven by existence of a file that won't be here when building
under mock.

References:
http://rpm.org/gitweb?p=rpm.git;a=commit;h=90463ac
https://github.com/rpm-software-management/mock/issues/11