Test: tools: make validity tests better adapting to new schemas
We only want to suppress boring repeated parts from the diff when
new schema emerges, but still want to be alerted about the new lines
arising from the presence of such new schema -- that's a feedback
loop that it is indeed being taken into account (and updating the
regression tests to reconcile this minimalistic diff is a bearable
duty at that opportunity).

In one go, also reflect latest addition (as of previous commit) to
the line of schemas -- pacemaker-2.7.