stonith: Check for missing params in new device and dup
device_has_duplicate would return a dup if all params in device where in dup with the same value but
if dup had a param that device did not the dup would still be considered a duplicate. This updates
the duplicate check to make sure that all keys and values overlap in both device and dup params.