Low: crmd: don't clear remote node transient attributes unnecessarily
cd10f0b modified crmd to update the CIB directly for remote node attributes
when legacy attrd was in use. This included clearing remote node transient
attributes when calling attrd_update_delegate() with 'C' (i.e.
ATTRD_OP_PEER_REMOVE). However, this is not necessary, because peer removal
is only intended to clears caches, not the CIB. The CIB is cleared separately,
e.g. in remote_node_up() and remote_node_down().

This also fixes a minor issue where an rc variable was mistyped and then
obscured by a redeclaration.