Feature: attrd: support Pacemaker Remote node attributes in legacy attrd
Previously, the crmd, crm_attribute and crm_resource would send
Pacemaker Remote node attribute updates directly to the CIB when legacy attrd
is in use. Now, legacy attrd implements the same logic, to simplify the
callers, and make future changes in resource failure handling easier.

This is not equivalent to tracking the attributes, as is done for cluster
nodes. In particular, dampening is ignored, and requests submitted to different
nodes cannot be ordered reliably.