Build: spec: %define back to %global to prevent recursive declaration
Commit 625d427 was too coarse and disregarded self-inflicted recursive
redefinition of an internal macro, owing to how %define is treated.
Unfortunately, rpmbuild won't currently abort the build in that case
even if it is aware of the recursive expansion [*], masking the problem.

The breakthrough emerged only thanks to materialized changes in the
resulting RPMs, specifically because __os_install_post macro definition
got suddently empty, disregarding common actions such as compressing
the man pages (beside various debug symbols stripping), which in turn
led to the discovery (good catch, Ken).

So simply revert back to using %global in this instance, which will
make it harmless one-off definition once again.  (And reformat to
spare overly long line.)

[*] https://github.com/rpm-software-management/rpm/issues/145