Refactor: lib/services: unify on single source of "synchronous" flag
Forerunner systemd.[ch] is followed by the rest of squad in decision
based merely on svc_action_t operand with suitably flipped member item
and no more on redundant extraneous parameter that in turn is dropped.
It was (for the former) and is possible because both a/synchronous paths
leading to such decision contains a priori setting of the flag as
expected (in services_action_async/services_action_sync, respectively).