Feature: attrd: add API for clearing failures
This will make future changes to failure handling easier.

An alternative would be to use the existing F_ATTRD_REGEX feature to clear
failures. The problem is that implementing regexes in legacy attrd when used
with remote node attributes would require reading all the attributes from the
CIB, then clearing each one that matches, which is race-prone and expensive.