Refactor: normalize header inclusion guards 1/2: existing ones
Originating in discussion around commit ef711f968, there's a Ken's
concern about not enough uniqueness in artificial header inclusion
guard names (in that particular case) on the table.

Together with varying styles used for these names across the project
and the guards sometimes missing altogether, this is to be addressed
in two subsequent commits (adjust existing + fix missing ones).

With this first step, we further introduce following naming scheme:

1. all guards start with PCMK ands end with __H

2. there are 5 purposeful patterns so that the guard name also encodes
   some not-so-obvious pieces of information:

   PCMKDEF_<HEADER>__H:
     shared headers being a mere dictionary assigning symbolic
     names to constants/literals, no type/function/object business here

   PCMKLIB_<LIBRARY-or-MIX>_<HEADER>__H:
     public headers for a library (or a mix of libraries designated
     with "MIX")

   PCMKPRIV_<LIBRARY-or-MIX>_<HEADER>__H:
     private headers for a library (or a mix of libraries designated
     with "MIX") created more-or-less in an organic way, function
     promises via "extern" and not via proper declarations, etc.

   PCMKINT_<LIBRARY-or-MIX>_<HEADER>__H:
     private headers for a library (or a mix of libraries designated
     with "MIX") created in more decent way, declaring functions
     properly (with hidden visibility, cf. G_GNUC_INTERNAL from glib)
     hence usable in all the contexts of standard library-internal use

   PCMK_<DIR-or-BINARY>_<HEADER>__H:
     private headers exclusively for compiling particular binary

Some kinds of normalization is also applied on lines around respective
preprocessor conditionals and comments.

(As an aside, there used to be unintendedly mixed "ElECTION" in a guard
name prior to this change.)